moved everything into a single file and made all the classes that need no exposing private to that static class that introduces the extension method on Uri
